pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "internvl-gradio-demo"
version = "0.1.0"
description = "Gradio demo for InternVL3.5 supporting VQA and text-guided object search"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Demo Maintainer" }]
dependencies = [
  "accelerate>=0.29.0",
  "gradio==4.44.1",
  "huggingface-hub>=0.23.0",
  "jinja2>=3.1.4",
  "numpy>=1.26.4",
  "opencv-python>=4.9.0",
  "pillow>=10.3.0",
  "pydantic>=2.7.1",
  "pydantic-settings>=2.0.0",
  "python-dotenv>=1.0.1",
  "safetensors>=0.4.3",
  "torch>=2.2.0",
  "torchvision>=0.17.0",
  "timm>=0.9.0",
  "einops>=0.7.0",
  "transformers>=4.40.0",
  "uvicorn>=0.30.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "ruff>=0.4.4"
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "I"]
ignore = ["E501"]
